+++ SHE_PPT/python/SHE_PPT/table_formats/she_psf_zm_state.py
@@ -23,12 +23,11 @@ __updated__ = "2021-08-13"
 # Boston, MA 02110-1301 USA
 
 from collections import OrderedDict
-from typing import Optional, List, Type
+from typing import List, Optional, Type
 
-from ..constants.fits import PSF_ZM_STATE_TAG, PSF_ZM_IDENTITY
+from ..constants.fits import PSF_ZM_IDENTITY, PSF_ZM_STATE_TAG
 from ..logging import getLogger
 from ..table_formats.she_psf_state import ShePsfStateFormat, ShePsfStateMeta
-
 
 fits_version = "8.0"
 
@@ -45,7 +44,7 @@ class ShePsfZmStateMeta(ShePsfStateMeta):
 
     def init_meta(self,
                   **kwargs: str) -> OrderedDict:
-        return super().init_meta(extname=PSF_ZM_STATE_TAG,
+        return super().init_meta(extname = PSF_ZM_STATE_TAG,
                                  **kwargs)
 
 
@@ -59,21 +58,20 @@ class ShePsfZmStateFormat(ShePsfStateFormat):
     _meta_type: Type = ShePsfZmStateMeta
     _l_colnames: Optional[List[str]] = None
 
-    def __init__(self, data_type="FIELD"):
-
+    def __init__(self, data_type = "FIELD"):
         super().__init__(data_type)
 
         # Column names and info
 
         self.fovrngx = self.set_column_properties(
-            "SHE_PSF_%s_FOVRNGX" % data_type, dtype=">f4",
-            fits_dtype="E", length=2)
+            "SHE_PSF_%s_FOVRNGX" % data_type, dtype = ">f4",
+            fits_dtype = "E", length = 2)
         self.fovrngy = self.set_column_properties(
-            "SHE_PSF_%s_FOVRNGY" % data_type, dtype=">f4",
-            fits_dtype="E", length=2)
+            "SHE_PSF_%s_FOVRNGY" % data_type, dtype = ">f4",
+            fits_dtype = "E", length = 2)
         self.zer_ply_amp = self.set_column_properties(
-            "SHE_PSF_%s_ZNKPLYAMP" % data_type, dtype=">f4",
-            fits_dtype="E", length=50)
+            "SHE_PSF_%s_ZNKPLYAMP" % data_type, dtype = ">f4",
+            fits_dtype = "E", length = 50)
 
         self._finalize_init()
